Revert "Seguimiento enemigo"
