--- _Project/Assets/2_Scripts/Enemy/EnemyAttackComponent.cs
@@ -1,28 +1,0 @@
-using UnityEngine;
-
-public class EnemyAttackComponent : MonoBehaviour, IAttackComponent
-{
-    
-
-    // Start is called once before the first execution of Update after the MonoBehaviour is created
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-    public void ActiveHitbox()
-    {
-        throw new System.NotImplementedException();
-    }
-
-    public void Attack()
-    {
-        throw new System.NotImplementedException();
-    }
-
-}
