--- _Project/Assets/2_Scripts/Enemy/EnemyAttackComponent.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 98bc7968b6dfbf543acedf409559a242
