--- _Project/Assets/2_Scripts/Enemy/EnemyComponents.asmdef
@@ -1,17 +1,0 @@
-{
-    "name": "EnemyComponents",
-    "rootNamespace": "",
-    "references": [
-        "GUID:8cb876c56ff9054478fa90b8c26431b8",
-        "GUID:b42da03a10049724b8c7d528d828690f"
-    ],
-    "includePlatforms": [],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": true,
-    "defineConstraints": [],
-    "versionDefines": [],
-    "noEngineReferences": false
-}
