--- _Project/Assets/2_Scripts/Enemy/EnemyComponents.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: cd211689d9bd069408dc0272187eab8d
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
