+++ _Project/Assets/2_Scripts/Enemy/EnemyController.cs
@@ -1,43 +1,16 @@
 using UnityEngine;
 
-public class EnemyController : MonoBehaviour, IStateMachineComponent
+public class EnemyController : MonoBehaviour
 {
-    [SerializeField] IStateComponent actualState;
-    [SerializeField] IStateComponent lastState;
-
+    [SerializeField] IDamageableComponent damageableComponent;
     void Awake()
     {
-        actualState = new GolemWanderState(this, transform);
-        actualState.Enter();
+        damageableComponent = GetComponent<IDamageableComponent>();
     }
 
     // Update is called once per frame
     void Update()
     {
-        MUpdate();
-    }
-
-    void FixedUpdate()
-    {
-        MFixedUpdate();
-    }
-
-    public void MUpdate()
-    {
-        actualState.Update();
-    }
-
-    public void MFixedUpdate()
-    {
-        actualState.FixedUpdate();
-    }
-
-    public void ChangeState(IStateComponent newState)
-    {
-        lastState = actualState;
-        actualState = newState;
-
-        lastState.Exit();
-        actualState.Enter();
+        
     }
 }
