--- _Project/Assets/2_Scripts/Enemy/EnemyMoveComponent.cs
@@ -1,26 +1,0 @@
-using UnityEngine;
-
-public class EnemyMoveComponent : MonoBehaviour, IMoveComponent
-{   
-    // Start is called once before the first execution of Update after the MonoBehaviour is created
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-
-    public void IsMoving(Vector2 m)
-    {
-        throw new System.NotImplementedException();
-    }
-
-    public void Move()
-    {
-        throw new System.NotImplementedException();
-    }
-}
