--- _Project/Assets/2_Scripts/Enemy/EnemyMoveComponent.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 93b6aadb5e739db4bb0856f59b2f6d03
