--- _Project/Assets/2_Scripts/Enemy/Golem.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: fee7d6b4f2dae1e40bdf20809095223f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
