--- _Project/Assets/2_Scripts/Enemy/Golem/GolemAttackState.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: d7aa629b5c2344049b492d1a79da5d2d
