--- _Project/Assets/2_Scripts/Enemy/Golem/GolemChaseState.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: a2f0242c6f0a47242ba0fdcb6e7553b3
