--- _Project/Assets/2_Scripts/Enemy/Golem/GolemWanderState.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 666ac3239e68aab4fb47b5093d12ab57
