--- _Project/Assets/2_Scripts/Interfaces.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9e3aa65b273bf964a94f1c96bff652c2
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
