--- _Project/Assets/2_Scripts/Interfaces/ISkillComponent.cs
@@ -1,7 +1,0 @@
-using UnityEngine;
-
-public interface ISkillComponent
-{
-    void InitializeSpecialSkill();
-    void DoSpecialSkill();
-}
