--- _Project/Assets/2_Scripts/Interfaces/ISkillComponent.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 1b1b3b55eb20de147827f70ad91d1abf
