--- _Project/Assets/2_Scripts/Interfaces/IStateComponent.cs
@@ -1,9 +1,0 @@
-using UnityEngine;
-
-public interface IStateComponent
-{
-    void Enter();
-    void Exit();
-    void Update();
-    void FixedUpdate();
-}
