--- _Project/Assets/2_Scripts/Interfaces/IStateComponent.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 594c5683930ecff449c4184ca42228e5
