--- _Project/Assets/2_Scripts/Interfaces/IStateMachineComponent.cs
@@ -1,8 +1,0 @@
-using UnityEngine;
-
-public interface IStateMachineComponent
-{
-    void MUpdate();
-    void MFixedUpdate();
-    void ChangeState(IStateComponent s);
-}
