--- _Project/Assets/2_Scripts/Interfaces/IStateMachineComponent.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 8e321faf577180940a5354e582f50a8c
