--- _Project/Assets/2_Scripts/Interfaces/Interfaces.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 8cb876c56ff9054478fa90b8c26431b8
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
