+++ _Project/Assets/2_Scripts/Player/IAttackComponent.cs
