+++ _Project/Assets/2_Scripts/Player/IMoveComponent.cs
@@ -2,6 +2,5 @@ using UnityEngine;
 using UnityEngine.InputSystem;
 public interface IMoveComponent
 {
-    void IsMoving(Vector2 m);
     void Move();
 }
