+++ _Project/Assets/2_Scripts/Player/PlayerComponents.asmdef
@@ -2,9 +2,7 @@
     "name": "PlayerComponents",
     "rootNamespace": "",
     "references": [
-        "GUID:75469ad4d38634e559750d17036d5f7c",
-        "GUID:8cb876c56ff9054478fa90b8c26431b8",
-        "GUID:b42da03a10049724b8c7d528d828690f"
+        "GUID:75469ad4d38634e559750d17036d5f7c"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
