--- _Project/Assets/2_Scripts/Utils.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f78477dfb04cafe46b72c5e0347898be
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
