--- _Project/Assets/2_Scripts/Utils/Utils.asmdef
@@ -1,3 +1,0 @@
-{
-	"name": "Utils"
-}
