--- _Project/Assets/2_Scripts/Utils/Utils.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: b42da03a10049724b8c7d528d828690f
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
