--- _Project/Assets/2_Scripts/Utils/VectorConverter.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: ab31308d9c35fa94db76b926546dba26
